Exceptions in Java
When reading the data input file, if there are not enough pieces of data
to create Student and StudentMarks objects then throw a
DataInputException.
The exception message should display how data pieces were provided vs.
how many are required.
This exception should be thrown by the file reading method.
The tricky part is to figure out how to count the pieces of data that
are read from the file.
4